TigerRag/tigerrag/requirements.txt
pandas
numpy
torch
transformers
faiss-cpu
scikit-learn
openai
